requirements.txt
# Requires Python >=3.12.2
requests==2.32.0
scapy==2.6.1
psutil==7.0.0
pysnmp==4.4.12
python-nmap==0.7.1
pyasn1==0.4.8
psycopg==3.2.9